simplify types
