+++ packages/graphql/schemas/schema.graphql
@@ -348,40 +348,10 @@ type DevState {
 type Editor {
   """Binary that opens the editor"""
   binary: String!
-  id: EditorId!
+  id: String!
 
   """name of editor"""
   name: String!
-}
-
-"""Editor id"""
-enum EditorId {
-  appcode
-  atom
-  atombeta
-  brackets
-  clion
-  clion64
-  code
-  computer
-  emacs
-  idea
-  idea64
-  insiders
-  notepad
-  other
-  phpstorm
-  phpstorm64
-  pycharm
-  pycharm64
-  rubymine
-  rubymine64
-  sublimetext
-  sublimetext2
-  sublimetextdev
-  vim
-  webstorm
-  webstorm64
 }
 
 """Represents a spec on the file system"""
