+++ packages/graphql/src/schemaTypes/objectTypes/gql-Editor.ts
@@ -1,19 +1,10 @@
-import { enumType, objectType } from 'nexus'
-import { editorIds } from '@packages/types'
-
-const EditorIdEnum = enumType({
-  name: 'EditorId',
-  description: 'Editor id',
-  members: editorIds,
-})
+import { objectType } from 'nexus'
 
 export const Editor = objectType({
   name: 'Editor',
   description: 'Represents an editor on the local machine',
   definition (t) {
-    t.nonNull.field('id', {
-      type: EditorIdEnum,
-    })
+    t.nonNull.string('id')
 
     t.nonNull.string('name', {
       description: 'name of editor',
